\name{CropPhenology-package}
\alias{CropPhenology-package}
\alias{CropPhenology}
\docType{package}
\title{
\packageTitle{CropPhenology}
}
\description{
\packageDescription{CropPhenology}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CropPhenology}
\packageIndices{CropPhenology}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CropPhenology}

Maintainer: \packageMaintainer{CropPhenology}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
