% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{MultiPointsPlot}
\alias{MultiPointsPlot}
\title{Time series curves for Multiple points in the Region of Interest}
\usage{
MultiPointsPlot(path, N, Id1, Id2, Id3, Id4, Id5)
}
\arguments{
\item{path}{- the path whee AllPixel.txt saved}

\item{N}{- number of intersted points}

\item{Id1}{-  ID number for point 1}

\item{Id2}{-  Id number for point 2}

\item{Id3}{-  ID number for point 3}

\item{Id4}{-  ID number for point 4}

\item{Id5}{-  ID number for point 5}
}
\value{
Multiple time series curves together at the plot panel
}
\description{
MultiPointsPlot function takes PixelIDs for more than one pixels within the region of interst and returns timeseries curves from these points, ploted together. The pixelIds scan be obtained from the txt file output of PhenoMetrics function, AllPixels.txt.
MultiPointsPlot function provides the ability to visualise the NDVI curve by plotting the temporal sequences of NDVI values of user selected raster pixels. It allows plots of a maximum of five pixels together. 
The resulting multiple curve plots allow assesment of the spatial and temporal crop growth variability among the selected locations, which provide important information about the environmental factors influencing crop growth (eg . Araya et al., 2016)
}

\examples{
MultiPointsPlot(system.file("extdata/data3", package="CropPhenology"),3,11,114,125)

}
\author{
Sofanit Araya
}
\references{
Araya, S., Lyle, G., Lewis, M., Ostendorf, B., 2016. Phenologic metrics derived from MODIS NDVI as indicators for Plant Available Water-holding Capacity.\emph{Ecological Indicators} \bold{60, 1263-1272.}
}

\seealso{
PhenoMetrics()
}
\keyword{Curve}
\keyword{curves}
\keyword{from}
\keyword{multiple}
\keyword{points}
\keyword{time-series}

