% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropPhenology.R
\name{PhenoMetrics}
\alias{PhenoMetrics}
\title{Phenologic metrics from time series vegetation index data}
\usage{
PhenoMetrics(Path, BolAOI, Percentage)
}
\arguments{
\item{Path}{- Text value - the path where the time series images saved}

\item{BolAOI}{- Logical value - if there is any area of intererst or not}

\item{Percentage}{- Optional Numeric Vlaue - percentage of minimum NDVI value at which the Onset and Offset is defined. The 'Percentage' paramenter is optional; if not provided, a Default value of 10 will be taken.}
}
\value{
OnsetV - NDVI value at the start of continuous positive slope between successive NDVI values above the user defined percentage threshold. It represent early gorth srages (seedling)

OnsetT - MODIS acquisition time when OnsetV is derived.

MaxV - the annual maximum NDVI, represents full canopy coverage during Anthesis stage

MaxT - the time when the maximum NDVI occurs, represents anthesis/flowering stage

OffsetV - NDVI value measured at the lowest slope below a user defined percentage threshold.

OffsetT - the time when the offsetV derived, represents time when crop has ripened

LengthGS- the length of the growing season between Onset and Offset

BeforeMaxT - The duration between the OnsetT and MaxT

AfterMaxT - The duration between the MaxT and OffsetT

GreenUpSlope- The rate at which NDVI increases from the OnsetV to MaxV over the time of the difference between MaxT and OnsetT

BrownDownSlope - The rate at which NDVI decreases from MaxV to OffsetV over the difference between OffsetT and MaxT

TINDVI - The sum of NDVI values attained at each image date within the growing season measured by the area under the NDVI curve.

TINDVIBeforeMax - the integral area under the curve between Onset and Maximum NDVI, indicates the pre-anthesis crop growth.

TINDVIAfterMax - the integral area under the curve between Maximum NDVI and Offset, indicates the post-anthesis growth.

Asymmetry - the difference between AreaBeforeMax and AreaAfterMax. It measures which part of the growing season attain relatively higher accumulated NDVI values
}
\description{
PhenoMetrics function extracts phenologic metrics from time series vegetaion index data, as raster and Ascii files. The function takes path of the vegetation index data, the boolean Value for BolAOI, and the percentage threshold value as a input parameters. 
The outputs metrics are plotted in R plot windows shown as maps within the R software for fast and easy visualisation within the R software.The metrics are represented spatially by fifteen raster images in the IMG format in the same projection as the input data which can be viewed from a geographic information system and are also exported as a generically formatted ASCII text file. Each of the ASCII files and the raster output metrics are named as the metrics name, for example OnsetT.txt and OnsetT.img. All output files will be then saved in a folder "Metrics" under the directory specified at the path parameter.} 


\examples{
# EXAMPLE - 1

PhenoMetrics(system.file("extdata/data1", package="CropPhenology"), FALSE)
 
# EXAMPLE - 2
 
PhenoMetrics(system.file("extdata/data2", package="CropPhenology"), TRUE, 15)


}
\author{
Sofanit Araya
}

\seealso{
MultiPointsPlot (Path, N,Id1, Id2...IdN)
}
\keyword{Phenology,}
\keyword{Time-series}
\keyword{image,}
\keyword{remote}
\keyword{satellite}
\keyword{sensing,}

