\name{CropPhenology-package}
\alias{CropPhenology-package}
\alias{CropPhenology}
\docType{package}
\title{
\packageTitle{CropPhenology}
}
\description{Phenologic data is a key environmental parameter to ecological studies. Time series data of remote sensing vegetation index have been long used to obtain the vegetation dynamics of crops (Henebry etal. 2013). One satellite sensor that has been extensively used to derived phenologic information is the Moderate Resolution Imaging Spectroradiometer (MODIS). This sensor has been used in a number of crop monitoring applications, including crop type mapping and classification (Zhong et al., 2011), regional yield estimation (Sakamoto et al., 2013), and yield forecasting (Bolton and Friedl, 2013). Its moderate resolution and high frequency of acquisition makes it particularly appropriate for characterising crop phenology. Several phenologic products have been developed at different spatial resolutions from the MODIS sensor. CropPhenology package extracts 15 phenologic metrics from MODIS time-series vegetation index data. It has minimum pre-processing of input data and easy interpreatable outputs. }

\author{
\packageAuthor{CropPhenology}

Maintainer: \packageMaintainer{CropPhenology}
}
\references{
Henebry, G., de Beurs, K., 2013. Remote Sensing of Land Surface Phenology: A Prospectus, in: Schwartz, M.D. (Ed.), Phenology: An Integrative Environmental Science. \emph Springer Netherlands, pp. 385-411.

Zhong, L., Hawkins, T., Biging, G., Gong, P., 2011. A phenology-based approach to map crop types in the San Joaquin Valley, California. \emph International Journal of Remote Sensing, \bold 32, 7777-7804.

Sakamoto, T., Gitelson, A.A., Arkebauer, T.J., 2013. MODIS-based corn grain yield estimation model incorporating crop phenology information.\emph Remote Sensing of Environment, \bold 131, 215-231.

Bolton, D.K., Friedl, M.A., 2013. Forecasting crop yield using remotely sensed vegetation indices and crop phenology metrics. \emph Agricultural and Forest Meteorology, \bold 173, 74-84.


}
\keyword{ package }
